requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
anyio=3.5.0=py310h06a4308_0
argon2-cffi=20.1.0=py310h7f8727e_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py310h06a4308_0
babel=2.11.0=py310h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310h06a4308_0
blas=1.1=openblas
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h9c3ff4c_4
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
cairo=1.16.0=hb05425b_5
certifi=2023.11.17=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
coloredlogs=15.0.1=pypi_0
comm=0.1.2=py310h06a4308_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=41.0.2=py310h774aba0_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl=1.1.1=py310_0
eigen=3.4.0=h4bd325d_0
entrypoints=0.4=py310h06a4308_0
et_xmlfile=1.1.0=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=py310h06a4308_0
flatbuffers=23.5.26=pypi_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
fsspec=2023.10.0=pyhca7485f_0
giflib=5.2.1=h36c2ea0_2
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.13=h85f3911_1
google-auth=2.25.2=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.60.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.12.1=nompi_h2750804_100
humanfriendly=10.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py310h2f386ee_0
ipython=8.8.0=py310h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=py310h06a4308_0
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h166bdaf_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py310h06a4308_0
jupyter=1.0.0=py310h06a4308_8
jupyter_client=7.4.8=py310h06a4308_0
jupyter_console=6.4.4=py310h06a4308_0
jupyter_core=5.1.1=py310h06a4308_0
jupyter_server=1.23.4=py310h06a4308_0
jupyterlab=3.5.3=py310h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.16.5=py310h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libclang=10.0.1=default_hb85057a_2
libcurl=8.2.1=h91b91d3_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=15_linux64_openblas
libllvm10=10.0.1=he513fc3_3
libnghttp2=1.52.0=ha637b67_1
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lightning-utilities=0.10.0=pyhd8ed1ab_0
lxml=4.9.1=py310h1edc446_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pypi_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib=3.7.2=py310hff52083_0
matplotlib-base=3.7.2=py310h1128e8f_0
matplotlib-inline=0.1.6=py310h06a4308_0
mistune=0.8.4=py310h7f8727e_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.4.8=py310h06a4308_0
nbclient=0.5.13=py310h06a4308_0
nbconvert=6.5.4=py310h06a4308_0
nbformat=5.7.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
nettle=3.6=he412f7d_0
networkx=3.1=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=6.5.2=py310h06a4308_0
notebook-shim=0.2.2=py310h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.4=py310h757a811_1
numpy=1.26.0=pypi_0
numpy-base=1.24.3=py310h375b286_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pyhd8ed1ab_0
onnxruntime=1.16.3=pypi_0
openblas=0.3.20=pthreads_h320a7e8_0
opencv=4.6.0=py310ha16400d_5
opencv-python=4.5.5.62=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.0.10=py310h5eee18b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.1.1=py310h1128e8f_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310h6a678d5_0
pip=22.3.1=py310h06a4308_0
pixman=0.40.0=h36c2ea0_0
platformdirs=2.5.2=py310h06a4308_0
ply=3.11=py_1
prometheus_client=0.14.1=py310h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py310h5eee18b_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.9=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=py310h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=2.0.1=cpu_py310hab5cca8_0
pytorch-lightning=2.1.1=pyhd8ed1ab_0
pytorch-mutex=1.0=cpu
pytz=2022.7=py310h06a4308_0
pyyaml=6.0=py310h5764c6d_4
pyzmq=23.2.0=py310h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.4.0=py310h06a4308_0
qtpy=2.2.0=py310h06a4308_0
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.31.0=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.0=py310h6a678d5_1
scipy=1.8.1=py310h7612f91_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.6.3=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py310h06a4308_1
soupsieve=2.3.2.post1=py310h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=pyh04b8f61_3
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.17.1=py310h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py310h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=py310h06a4308_0
torchmetrics=1.2.0=pyhd8ed1ab_0
torchvision=0.15.2=cpu_py310h83e0c9b_0
tornado=6.3.2=py310h5eee18b_0
tqdm=4.65.0=py310h2f386ee_0
traitlets=5.7.1=py310h06a4308_0
typing-extensions=4.7.1=py310h06a4308_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.16=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310h06a4308_1
websocket-client=0.58.0=py310h06a4308_4
werkzeug=3.0.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
